Smaller RawTick: redefine source_offset to be u32
This improves performance by about 8%.

+++ src/backend/errors/backend_error_kind.rs
@@ -14,7 +14,7 @@ pub enum BackendErrorKind<'a> {
         multichar: (char, u8),
         /// something else
         invalid: (char, TabElement),
-        tick_idx: usize,
+        tick_idx: u32,
     },
     InvalidPartlineSyntax(&'a str),
     // TODO: a parser error for invalid string names
